+++ src/graph/pages/mod.rs
@@ -9,3 +9,4 @@ pub mod media_upload;
 pub mod photo;
 pub mod utils;
 pub mod video;
+pub mod form_data;
